// DXC_Frontend/src/api/models/userDto.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * DXC_Core.API
 * OpenAPI spec version: 1.0
 */

export interface UserDto {
  publicId?: string
  /** @nullable */
  fullName: string | null
  /** @nullable */
  userName: string | null
  /** @nullable */
  email: string | null
  isActive?: boolean
  createdAt?: string
  /** @nullable */
  roleCodes?: string[] | null
  /** @nullable */
  departmentPublicId?: string | null
  /** @nullable */
  departmentName?: string | null
  /** @nullable */
  organizationPublicId?: string | null
  /** @nullable */
  organizationName?: string | null
}